Fix stack experience split build breakages on iOS
