+++ client/windows/CStackExperienceDetailsWindow.cpp
@@ -4,6 +4,33 @@
 #include "StdInc.h"
 #include "CStackExperienceDetailsWindow.h"
 
+#include <vcmi/spells/Spell.h>
+
+#include "../CPlayerInterface.h"
+#include "../GameEngine.h"
+#include "../GameInstance.h"
+#include "../gui/Shortcut.h"
+#include "../gui/WindowHandler.h"
+#include "../render/CAnimation.h"
+#include "../render/IImage.h"
+#include "../render/IRenderHandler.h"
+#include "../render/ImageLocator.h"
+#include "../widgets/Buttons.h"
+#include "../widgets/GraphicalPrimitiveCanvas.h"
+#include "../widgets/Images.h"
+#include "../widgets/ObjectLists.h"
+#include "../widgets/Slider.h"
+#include "../widgets/TextControls.h"
+
+#include "../../lib/CBonusTypeHandler.h"
+#include "../../lib/CStack.h"
+#include "../../lib/GameLibrary.h"
+#include "../../lib/bonuses/Propagators.h"
+#include "../../lib/callback/CCallback.h"
+#include "../../lib/entities/hero/CHeroHandler.h"
+#include "../../lib/texts/CGeneralTextHandler.h"
+#include "../../lib/texts/TextOperations.h"
+
 int CStackWindow::StackExperienceDetailsWindow::getStackExperienceTierFromCreatureLevel(int creatureLevel)
 {
 	const int maxTier = static_cast<int>(LIBRARY->creh->expRanks.size()) - 1;
@@ -568,517 +595,3 @@ void CStackWindow::StackExperienceDetailsWindow::rebuildTableRows()
 
 class CCreatureArtifactInstance;
 class CSelectableSkill;
-
-class UnitView
-{
-public:
-	// helper structs
-	struct CommanderLevelInfo
-	{
-		std::vector<ui32> skills;
-		std::function<void(ui32)> callback;
-	};
-	struct StackDismissInfo
-	{
-		std::function<void()> callback;
-	};
-	struct StackUpgradeInfo
-	{
-		StackUpgradeInfo() = delete;
-		StackUpgradeInfo(const UpgradeInfo & upgradeInfo)
-			: info(upgradeInfo)
-		{ }
-		UpgradeInfo info;
-		std::function<void(CreatureID)> callback;
-	};
-
-	// pointers to permanent objects in game state
-	const CCreature * creature;
-	const CCommanderInstance * commander;
-	const CStackInstance * stackNode;
-	const CStack * stack;
-	const CGHeroInstance * owner;
-
-	// temporary objects which should be kept as copy if needed
-	std::optional<CommanderLevelInfo> levelupInfo;
-	std::optional<StackDismissInfo> dismissInfo;
-	std::optional<StackUpgradeInfo> upgradeInfo;
-
-	// misc fields
-	unsigned int creatureCount;
-	bool popupWindow;
-
-	UnitView()
-		: creature(nullptr),
-		commander(nullptr),
-		stackNode(nullptr),
-		stack(nullptr),
-		owner(nullptr),
-		creatureCount(0),
-		popupWindow(false)
-	{
-	}
-
-	std::string getName() const
-	{
-		if(commander)
-			return commander->getType()->getNameSingularTranslated();
-		else
-			return creature->getNamePluralTranslated();
-	}
-private:
-
-};
-
-CCommanderSkillIcon::CCommanderSkillIcon(std::shared_ptr<CIntObject> object_, bool isMasterAbility_, std::function<void()> callback)
-	: object(),
-	  isMasterAbility(isMasterAbility_),
-	  isSelected(false),
-	  callback(callback)
-{
-	pos = object_->pos;
-	this->isMasterAbility = isMasterAbility_;
-	setObject(object_);
-}
-
-void CCommanderSkillIcon::setObject(std::shared_ptr<CIntObject> newObject)
-{
-	if(object)
-		removeChild(object.get());
-	object = newObject;
-	addChild(object.get());
-	object->moveTo(pos.topLeft());
-	redraw();
-}
-
-void CCommanderSkillIcon::clickPressed(const Point & cursorPosition)
-{
-	callback();
-	isSelected = true;
-	redraw();
-}
-
-void CCommanderSkillIcon::deselect()
-{
-	isSelected = false;
-	redraw();
-}
-
-bool CCommanderSkillIcon::getIsMasterAbility()
-{
-	return isMasterAbility;
-}
-
-void CCommanderSkillIcon::show(Canvas &to)
-{
-	CIntObject::show(to);
-
-	if(isMasterAbility && isSelected)
-		to.drawBorder(pos, Colors::YELLOW, 2);
-}
-
-static ImagePath skillToFile(int skill, int level, bool selected)
-{
-	// FIXME: is this a correct handling?
-	// level 0 = skill not present, use image with "no" suffix
-	// level 1-5 = skill available, mapped to images indexed as 0-4
-	// selecting skill means that it will appear one level higher (as if already upgraded)
-	std::string file = "zvs/Lib1.res/_";
-	switch (skill)
-	{
-		case ECommander::ATTACK:
-			file += "AT";
-			break;
-		case ECommander::DEFENSE:
-			file += "DF";
-			break;
-		case ECommander::HEALTH:
-			file += "HP";
-			break;
-		case ECommander::DAMAGE:
-			file += "DM";
-			break;
-		case ECommander::SPEED:
-			file += "SP";
-			break;
-		case ECommander::SPELL_POWER:
-			file += "MP";
-			break;
-	}
-	std::string suffix;
-	if (selected)
-		level++; // UI will display resulting level
-	if (level == 0)
-		suffix = "no"; //not available - no number
-	else
-		suffix = std::to_string(level-1);
-	if (selected)
-		suffix += "="; //level-up highlight
-
-	return ImagePath::builtin(file + suffix + ".bmp");
-}
-
-CStackWindow::CWindowSection::CWindowSection(CStackWindow * parent, const ImagePath & backgroundPath, int yOffset)
-	: parent(parent)
-{
-	pos.y += yOffset;
-	OBJECT_CONSTRUCTION;
-	if(!backgroundPath.empty())
-	{
-		background = std::make_shared<CPicture>(backgroundPath);
-		pos.w = background->pos.w;
-		pos.h = background->pos.h;
-	}
-}
-
-CStackWindow::ActiveSpellsSection::ActiveSpellsSection(CStackWindow * owner, int yOffset)
-	: CWindowSection(owner, ImagePath::builtin("stackWindow/spell-effects"), yOffset)
-{
-	static const Point firstPos(6, 2); // position of 1st spell box
-	static const Point offset(54, 0);  // offset of each spell box from previous
-
-	OBJECT_CONSTRUCTION;
-
-	const CStack * battleStack = parent->info->stack;
-
-	assert(battleStack); // Section should be created only for battles
-
-	//spell effects
-	int printed=0; //how many effect pics have been printed
-	std::vector<SpellID> spells = battleStack->activeSpells();
-	for(SpellID effect : spells)
-	{
-		const spells::Spell * spell = LIBRARY->spells()->getById(effect);
-
-		//not all effects have graphics (for eg. Acid Breath)
-		//for modded spells iconEffect is added to SpellInt.def
-		const bool hasGraphics = (effect < SpellID::THUNDERBOLT) || (effect >= SpellID::AFTER_LAST);
-
-		if (hasGraphics)
-		{
-			auto spellBonuses = battleStack->getBonuses(Selector::source(BonusSource::SPELL_EFFECT, BonusSourceID(effect)));
-			if (spellBonuses->empty())
-				throw std::runtime_error("Failed to find effects for spell " + effect.toSpell()->getJsonKey());
-
-			int duration = spellBonuses->front()->turnsRemain;
-			std::string preferredLanguage = LIBRARY->generaltexth->getPreferredLanguage();
-
-			MetaString spellText;
-			spellText.appendTextID(spell->getDescriptionTextID(0)); // TODO: select correct mastery level?
-			spellText.appendRawString("\n");
-			spellText.appendTextID(Languages::getPluralFormTextID( preferredLanguage, duration, "vcmi.battleResultsWindow.spellDurationRemaining"));
-			spellText.replaceNumber(duration);
-			std::string spellDescription = spellText.toString();
-
-			spellIcons.push_back(std::make_shared<CAnimImage>(AnimationPath::builtin("SpellInt"), effect + 1, 0, firstPos.x + offset.x * printed, firstPos.y + offset.y * printed));
-			labels.push_back(std::make_shared<CLabel>(firstPos.x + offset.x * printed + 46, firstPos.y + offset.y * printed + 36, EFonts::FONT_TINY, ETextAlignment::BOTTOMRIGHT, Colors::WHITE, std::to_string(duration)));
-			clickableAreas.push_back(std::make_shared<LRClickableAreaWText>(Rect(firstPos + offset * printed, Point(50, 38)), spellDescription, spellDescription));
-			if(++printed >= 8) // interface limit reached
-				break;
-		}
-	}
-}
-
-CStackWindow::BonusLineSection::BonusLineSection(CStackWindow * owner, size_t lineIndex)
-	: CWindowSection(owner, ImagePath::builtin("stackWindow/bonus-effects"), 0)
-{
-	OBJECT_CONSTRUCTION;
-
-	static const std::array<Point, 2> offset =
-	{
-		Point(6, 2),
-		Point(214, 2)
-	};
-
-	auto drawBonusSource = [this](int leftRight, Point p, BonusInfo & bi)
-	{
-		std::map<BonusSource, ColorRGBA> bonusColors = {
-			{BonusSource::ARTIFACT,          Colors::GREEN},
-			{BonusSource::ARTIFACT_INSTANCE, Colors::GREEN},
-			{BonusSource::CREATURE_ABILITY,  Colors::YELLOW},
-			{BonusSource::SPELL_EFFECT,      Colors::ORANGE},
-			{BonusSource::SECONDARY_SKILL,   Colors::PURPLE},
-			{BonusSource::HERO_SPECIAL,      Colors::PURPLE},
-			{BonusSource::STACK_EXPERIENCE,  Colors::CYAN},
-			{BonusSource::COMMANDER,         Colors::CYAN},
-		};
-
-		std::map<BonusSource, std::string> bonusNames = {
-			{BonusSource::ARTIFACT,          LIBRARY->generaltexth->translate("vcmi.bonusSource.artifact")},
-			{BonusSource::ARTIFACT_INSTANCE, LIBRARY->generaltexth->translate("vcmi.bonusSource.artifact")},
-			{BonusSource::CREATURE_ABILITY,  LIBRARY->generaltexth->translate("vcmi.bonusSource.creature")},
-			{BonusSource::SPELL_EFFECT,      LIBRARY->generaltexth->translate("vcmi.bonusSource.spell")},
-			{BonusSource::SECONDARY_SKILL,   LIBRARY->generaltexth->translate("vcmi.bonusSource.hero")},
-			{BonusSource::HERO_SPECIAL,      LIBRARY->generaltexth->translate("vcmi.bonusSource.hero")},
-			{BonusSource::STACK_EXPERIENCE,  LIBRARY->generaltexth->translate("vcmi.bonusSource.commander")},
-			{BonusSource::COMMANDER,         LIBRARY->generaltexth->translate("vcmi.bonusSource.commander")},
-		};
-
-		auto c = bonusColors.count(bi.bonusSource) ? bonusColors[bi.bonusSource] : ColorRGBA(192, 192, 192);
-		std::string t = bonusNames.count(bi.bonusSource) ? bonusNames[bi.bonusSource] : LIBRARY->generaltexth->translate("vcmi.bonusSource.other");
-		int maxLen = 50;
-		EFonts f = FONT_TINY;
-		Point pText = p + Point(4, 38);
-
-		// 1px Black border
-		bonusSource[leftRight].push_back(std::make_shared<CLabel>(pText.x - 1, pText.y, f, ETextAlignment::TOPLEFT, Colors::BLACK, t, maxLen));
-		bonusSource[leftRight].push_back(std::make_shared<CLabel>(pText.x + 1, pText.y, f, ETextAlignment::TOPLEFT, Colors::BLACK, t, maxLen));
-		bonusSource[leftRight].push_back(std::make_shared<CLabel>(pText.x, pText.y - 1, f, ETextAlignment::TOPLEFT, Colors::BLACK, t, maxLen));
-		bonusSource[leftRight].push_back(std::make_shared<CLabel>(pText.x, pText.y + 1, f, ETextAlignment::TOPLEFT, Colors::BLACK, t, maxLen));
-		bonusSource[leftRight].push_back(std::make_shared<CLabel>(pText.x, pText.y, f, ETextAlignment::TOPLEFT, c, t, maxLen));
-
-		frame[leftRight] = std::make_shared<GraphicalPrimitiveCanvas>(Rect(p.x, p.y, 52, 52));
-		frame[leftRight]->addRectangle(Point(0, 0), Point(52, 52), c);
-	};
-
-	for(size_t leftRight : {0, 1})
-	{
-		auto position = offset[leftRight];
-		size_t bonusIndex = lineIndex * 2 + leftRight;
-
-		if(parent->activeBonuses.size() > bonusIndex)
-		{
-			BonusInfo & bi = parent->activeBonuses[bonusIndex];
-			if (!bi.imagePath.empty())
-				icon[leftRight] = std::make_shared<CPicture>(bi.imagePath, position.x, position.y);
-
-			description[leftRight] = std::make_shared<CMultiLineLabel>(Rect(position.x + 60, position.y, 137, 50), FONT_TINY, ETextAlignment::TOPLEFT, Colors::WHITE, bi.description);
-			drawBonusSource(leftRight, Point(position.x - 1, position.y - 1), bi);
-		}
-	}
-}
-
-CStackWindow::BonusesSection::BonusesSection(CStackWindow * owner, int yOffset, std::optional<size_t> preferredSize):
-	CWindowSection(owner, {}, yOffset)
-{
-	OBJECT_CONSTRUCTION;
-
-	// size of single image for an item
-	static const int itemHeight = 59;
-
-	size_t totalSize = (owner->activeBonuses.size() + 1) / 2;
-	size_t visibleSize = preferredSize.value_or(std::min<size_t>(3, totalSize));
-
-	pos.w = owner->pos.w;
-	pos.h = itemHeight * (int)visibleSize;
-
-	auto onCreate = [=](size_t index) -> std::shared_ptr<CIntObject>
-	{
-		return std::make_shared<BonusLineSection>(owner, index);
-	};
-
-	lines = std::make_shared<CListBox>(onCreate, Point(0, 0), Point(0, itemHeight), visibleSize, totalSize, 0, totalSize > 3 ? 1 : 0, Rect(pos.w - 15, 0, pos.h, pos.h));
-	lines->onScroll = [owner](){ owner->redraw(); };
-}
-
-CStackWindow::ButtonsSection::ButtonsSection(CStackWindow * owner, int yOffset)
-	: CWindowSection(owner, ImagePath::builtin("stackWindow/button-panel"), yOffset)
-{
-	OBJECT_CONSTRUCTION;
-
-	if(parent->info->dismissInfo && parent->info->dismissInfo->callback)
-	{
-		auto onDismiss = [this]()
-		{
-			parent->info->dismissInfo->callback();
-			parent->close();
-		};
-		auto onClick = [=] ()
-		{
-			GAME->interface()->showYesNoDialog(LIBRARY->generaltexth->allTexts[12], onDismiss, nullptr);
-		};
-		dismiss = std::make_shared<CButton>(Point(5, 5),AnimationPath::builtin("IVIEWCR2.DEF"), LIBRARY->generaltexth->zelp[445], onClick, EShortcut::HERO_DISMISS);
-	}
-
-	if(parent->info->upgradeInfo && !parent->info->commander)
-	{
-		// used space overlaps with commander switch button
-		// besides - should commander really be upgradeable?
-
-		auto & upgradeInfo = parent->info->upgradeInfo.value();
-		const size_t buttonsToCreate = std::min<size_t>(upgradeInfo.info.size(), upgrade.size());
-
-		for(size_t buttonIndex = 0; buttonIndex < buttonsToCreate; buttonIndex++)
-		{
-			TResources totalCost = upgradeInfo.info.getAvailableUpgradeCosts().at(buttonIndex) * parent->info->creatureCount;
-
-			auto onUpgrade = [this, upgradeInfo, buttonIndex]()
-			{
-				upgradeInfo.callback(upgradeInfo.info.getAvailableUpgrades().at(buttonIndex));
-				parent->close();
-			};
-			auto onClick = [totalCost, onUpgrade]()
-			{
-				std::vector<std::shared_ptr<CComponent>> resComps;
-				for(TResources::nziterator i(totalCost); i.valid(); i++)
-				{
-					resComps.push_back(std::make_shared<CComponent>(ComponentType::RESOURCE, i->resType, i->resVal));
-				}
-
-				if(GAME->interface()->cb->getResourceAmount().canAfford(totalCost))
-				{
-					GAME->interface()->showYesNoDialog(LIBRARY->generaltexth->allTexts[207], onUpgrade, nullptr, resComps);
-				}
-					else
-					{
-						GAME->interface()->showInfoDialog(LIBRARY->generaltexth->allTexts[314], resComps);
-					}
-			};
-			auto upgradeBtn = std::make_shared<CButton>(Point(221 + (int)buttonIndex * 40, 5), AnimationPath::builtin("stackWindow/upgradeButton"), LIBRARY->generaltexth->zelp[446], onClick);
-
-			upgradeBtn->setOverlay(std::make_shared<CAnimImage>(AnimationPath::builtin("CPRSMALL"), LIBRARY->creh->objects[upgradeInfo.info.getAvailableUpgrades()[buttonIndex]]->getIconIndex()));
-
-			if(buttonsToCreate == 1) // single upgrade available
-				upgradeBtn->assignedKey = EShortcut::RECRUITMENT_UPGRADE;
-
-			upgrade[buttonIndex] = upgradeBtn;
-		}
-	}
-
-	if(parent->info->commander)
-	{
-		for(size_t buttonIndex = 0; buttonIndex < 2; buttonIndex++)
-		{
-			std::string btnIDs[2] = { "showSkills", "showBonuses" };
-			auto onSwitch = [buttonIndex, this]()
-			{
-				logAnim->debug("Switch %d->%d", parent->activeTab, buttonIndex);
-
-				parent->switchButtons[parent->activeTab]->enable();
-				parent->commanderTab->setActive(buttonIndex);
-				parent->switchButtons[buttonIndex]->disable();
-				parent->redraw(); // FIXME: enable/disable don't redraw screen themselves
-			};
-
-			std::string tooltipText = "vcmi.creatureWindow." + btnIDs[buttonIndex];
-			parent->switchButtons[buttonIndex] = std::make_shared<CButton>(Point(342, 5), AnimationPath::builtin("stackWindow/upgradeButton"), CButton::tooltipLocalized(tooltipText), onSwitch);
-			parent->switchButtons[buttonIndex]->setOverlay(std::make_shared<CAnimImage>(AnimationPath::builtin("stackWindow/switchModeIcons"), buttonIndex));
-		}
-		parent->switchButtons[parent->activeTab]->disable();
-	}
-
-	exit = std::make_shared<CButton>(Point(382, 5), AnimationPath::builtin("hsbtns.def"), LIBRARY->generaltexth->zelp[447], [this](){ parent->close(); }, EShortcut::GLOBAL_RETURN);
-}
-
-CStackWindow::CommanderMainSection::CommanderMainSection(CStackWindow * owner, int yOffset)
-	: CWindowSection(owner, ImagePath::builtin("stackWindow/commander-bg"), yOffset)
-{
-	OBJECT_CONSTRUCTION;
-
-	auto getSkillPos = [](int index)
-	{
-		return Point(10 + 80 * (index%3), 20 + 80 * (index/3));
-	};
-
-	auto getSkillImage = [this](int skillIndex)
-	{
-		bool selected = ((parent->selectedSkill == skillIndex) && parent->info->levelupInfo );
-		return skillToFile(skillIndex, parent->info->commander->secondarySkills[skillIndex], selected);
-	};
-
-	auto getSkillDescription = [this](int skillIndex) -> std::string
-	{
-		return parent->getCommanderSkillDescription(skillIndex, parent->info->commander->secondarySkills[skillIndex]);
-	};
-
-	for(int index = ECommander::ATTACK; index <= ECommander::SPELL_POWER; ++index)
-	{
-		Point skillPos = getSkillPos(index);
-
-		auto icon = std::make_shared<CCommanderSkillIcon>(std::make_shared<CPicture>(getSkillImage(index), skillPos.x, skillPos.y), false, [=]()
-		{
-			GAME->interface()->showInfoDialog(getSkillDescription(index));
-		});
-
-		icon->text = getSkillDescription(index); //used to handle right click description via LRClickableAreaWText::ClickRight()
-
-		if(parent->selectedSkill == index)
-			parent->selectedIcon = icon;
-
-		if(parent->info->levelupInfo && vstd::contains(parent->info->levelupInfo->skills, index)) // can be upgraded - enable selection switch
-		{
-			if(parent->selectedSkill == index)
-				parent->setSelection(index, icon);
-
-			icon->callback = [this, index, icon]()
-			{
-				parent->setSelection(index, icon);
-			};
-		}
-
-		skillIcons.push_back(icon);
-	}
-
-	auto getArtifactPos = [](int index)
-	{
-		return Point(269 + 52 * (index % 3), 22 + 52 * (index / 3));
-	};
-
-	for(auto equippedArtifact : parent->info->commander->artifactsWorn)
-	{
-		Point artPos = getArtifactPos(equippedArtifact.first);
-		const auto commanderArt = equippedArtifact.second.getArt();
-		assert(commanderArt);
-		auto artPlace = std::make_shared<CCommanderArtPlace>(artPos, parent->info->owner, equippedArtifact.first, commanderArt->getTypeId());
-		artifacts.push_back(artPlace);
-	}
-
-	if(parent->info->levelupInfo)
-	{
-		static constexpr ui32 commanderAbilitySkillOffset = 100;
-
-		abilitiesBackground = std::make_shared<CPicture>(ImagePath::builtin("stackWindow/commander-abilities.png"));
-		abilitiesBackground->moveBy(Point(0, pos.h));
-
-		std::vector<ui32> abilitySkills;
-		abilitySkills.reserve(parent->info->levelupInfo->skills.size());
-		for(ui32 skillID : parent->info->levelupInfo->skills)
-		{
-			if(skillID >= commanderAbilitySkillOffset)
-				abilitySkills.push_back(skillID);
-		}
-		size_t abilitiesCount = abilitySkills.size();
-
-		auto onCreate = [this, abilitySkills](size_t index)->std::shared_ptr<CIntObject>
-		{
-			if(index >= abilitySkills.size())
-				return nullptr;
-
-			const ui32 skillID = abilitySkills[index];
-			const auto bonuses = LIBRARY->creh->skillRequirements[skillID - commanderAbilitySkillOffset].first;
-			const CStackInstance * stack = parent->info->commander;
-			auto icon = std::make_shared<CCommanderSkillIcon>(std::make_shared<CPicture>(stack->bonusToGraphics(bonuses[0])), true, [](){});
-			icon->callback = [this, skillID, icon]()
-			{
-				parent->setSelection(skillID, icon);
-			};
-			std::string abilityDescription;
-			for(size_t i = 0; i < bonuses.size(); i++)
-			{
-				if(!abilityDescription.empty())
-					abilityDescription += "\n";
-
-				abilityDescription += LIBRARY->bth->bonusToString(bonuses[i]);
-			}
-
-			icon->hoverText = abilityDescription;
-			icon->text = abilityDescription;
-
-			return icon;
-		};
-
-		abilities = std::make_shared<CListBox>(onCreate, Point(38, 3+pos.h), Point(63, 0), 6, abilitiesCount);
-		abilities->onScroll = [owner](){ owner->redraw(); };
-
-		leftBtn = std::make_shared<CButton>(Point(10,  pos.h + 6), AnimationPath::builtin("hsbtns3.def"), CButton::tooltip(), [this](){ abilities->moveToPrev(); }, EShortcut::MOVE_LEFT);
-		rightBtn = std::make_shared<CButton>(Point(411, pos.h + 6), AnimationPath::builtin("hsbtns5.def"), CButton::tooltip(), [this](){ abilities->moveToNext(); }, EShortcut::MOVE_RIGHT);
-
-		if(abilitiesCount <= 6)
-		{
-			leftBtn->block(true);
-			rightBtn->block(true);
-		}
-
-		pos.h += abilitiesBackground->pos.h;
-	}
-}
